// apps/web/src/lib/types.ts
export interface NavItem {
  label: string;
  href: string;
  children?: Array<{
    label: string;
    description: string;
    href: string;
  }>;
}
// ...
export interface SocialLink {
  label: string;
  href: string;
}